premierleague/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2 
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncio==3.4.3
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
bs4==0.0.2
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fluent-logger==0.11.1
fqdn==1.5.1
frozenlist==1.6.0
google-auth==2.38.0
google-genai==1.0.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
msgpack==1.1.1
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.3.0
openai==1.61.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
playwright==1.52.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.1
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.18.0
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
selenium==4.27.1
Send2Trash==1.8.3
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
syncio==0.0.4
terminado==0.18.1
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.20.0